+++ changelog.md
@@ -68,8 +68,9 @@
 - Fixed UI a bit to fit the window a bit better
 
 # 3.1 - Emergency Bugfix and what have you
-- Fixed MyShop Portrait Display to be 135x135--proper MyShop resolution
 - Added additional user control over BMP output regarding background
+- Added palette export options in Custom Preview Options Menu for more artist colorpicking ease
+- Fixed MyShop Portrait Display to be 135x135--proper MyShop resolution
 - Fine-tuned the new UI on the Custom Preview Options Menu
 - Fixed Custom Frame Display being finnicky with numbers (the numbers, Mason, what do they mean?!?!?)
 - Fixed User Chosen Frame Export being turned on by default and causing problems